+++ packages/ui/icons/package.json
@@ -7,7 +7,6 @@
     "lint": "eslint \"**/*.{js,jsx,cjs,mjs,ts,tsx,cts,mts}\""
   },
   "dependencies": {
-    "@favolink-ui/system": "workspace:*",
     "@favolink-ui/utils": "workspace:*",
     "react": "^18.2.0"
   },
