+++ packages/ui/icons/src/icon/create-Icon.tsx
@@ -1,6 +1,6 @@
-import { forwardRef } from '@favolink-ui/system';
 import { Children, type ReactNode } from 'react';
 import { Icon, type IconProps } from './icon';
+import { forwardRef } from '../../../react/src/system';
 
 type CreateIconOptions = {
   d?: string;
