+++ packages/ui/react/src/components/button/icon-button.tsx
@@ -1,11 +1,7 @@
-import {
-  type HTMLFavolinkProps,
-  favolink,
-  forwardRef,
-} from '@favolink-ui/system';
 import { cx } from '@favolink-ui/utils';
 import { type ReactElement } from 'react';
 import * as styles from './icon-button.styles.css';
+import { type HTMLFavolinkProps, favolink, forwardRef } from '../../system';
 
 export type IconButtonProps = Omit<HTMLFavolinkProps<'button'>, 'children'> &
   styles.IconButtonVariants & {
