+++ packages/ui/react/src/components/menu/menu.context.ts
@@ -1,5 +1,5 @@
-import { createContext } from '@favolink-ui/system';
 import { type Ref } from 'react';
+import { createContext } from '../../system';
 
 type MenuContextValue = {
   open: boolean;
