+++ packages/ui/react/src/components/tag/tag.tsx
@@ -1,11 +1,7 @@
-import {
-  type HTMLFavolinkProps,
-  favolink,
-  forwardRef,
-} from '@favolink-ui/system';
 import { cx } from '@favolink-ui/utils';
 import * as styles from './tag.css';
 import { type MarginVariants, extractMarginProps } from '../../margin';
+import { type HTMLFavolinkProps, favolink, forwardRef } from '../../system';
 
 export type TagProps = HTMLFavolinkProps<'span'> &
   MarginVariants &
