+++ packages/ui/react/src/index.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @stylistic/padding-line-between-statements, react-refresh/only-export-components */
 export * from '@favolink-ui/hooks';
 export * from '@favolink-ui/icons';
-export * from '@favolink-ui/system';
+export * from './system';
 export * from './styles';
 export * from '@favolink-ui/utils';
 export * from './components';
